Add 게시글 삭제 기능

+++ movie_api/boards/views.py
@@ -46,9 +46,13 @@ def detail(request, article_pk):
         comment_serializer = CommentListSerializer(comments, many=True)
         return Response((article_serializer.data, comment_serializer.data))
     elif request.method == "DELETE" :
+        if not request.user == article.user:
+            return Response({' 삭제할 권한이 없습니다. '})
         article.delete()
         return Response({' 성공적으로 삭제되었습니다.'})
     elif request.method == "PUT" :
+        if not request.user == article.user:
+            return Response({ '수정할 권한이 없습니다. '})
         serializer = ArticleSerializer(article, data = request.data)
         if serializer.is_valid():
             serializer.save()
